Move the base verify() implementation into a mixin class
