+++ plumbum/machines/local.py
@@ -17,6 +17,7 @@ from plumbum.lib import ProcInfo, IS_WIN32, six, StaticProperty
 from plumbum.commands.daemons import win32_daemonize, posix_daemonize
 from plumbum.commands.processes import iter_lines
 from plumbum.machines.base import BaseMachine
+from plumbum.machines.base import PopenAddons
 from plumbum.machines.env import BaseEnv
 
 if sys.version_info >= (3, 2):
@@ -32,24 +33,11 @@ else:
         from subprocess import Popen, PIPE
         has_new_subprocess = False
 
-class IterablePopen(Popen):
+class IterablePopen(Popen, PopenAddons):
     iter_lines = iter_lines
     def __iter__(self):
         return self.iter_lines()
 
-    def verify(self, retcode, timeout, stdout, stderr):
-        if getattr(self, "_timed_out", False):
-            raise ProcessTimedOut("Process did not terminate within %s seconds" % (timeout,),
-                getattr(self, "argv", None))
-
-        if retcode is not None:
-            if hasattr(retcode, "__contains__"):
-                if self.returncode not in retcode:
-                    raise ProcessExecutionError(getattr(self, "argv", None), self.returncode,
-                        stdout, stderr)
-            elif self.returncode != retcode:
-                raise ProcessExecutionError(getattr(self, "argv", None), self.returncode,
-                    stdout, stderr)
 
 logger = logging.getLogger("plumbum.local")
